+++ src/main/php/io/streams/FileOutputStream.class.php
@@ -7,7 +7,7 @@ use io\File;
  *
  * @test  xp://net.xp_framework.unittest.io.streams.FileOutputStreamTest
  */
-class FileOutputStream implements OutputStream, Truncation {
+class FileOutputStream implements OutputStream, Seekable, Truncation {
   protected $file;
   
   /**
@@ -30,6 +30,19 @@ class FileOutputStream implements OutputStream, Truncation {
     $this->file->write($arg);
   }
 
+  /**
+   * Seek to a given offset
+   *
+   * @param   int $offset
+   * @param   int $whence default SEEK_SET (one of SEEK_[SET|CUR|END])
+   * @throws  io.IOException in case of error
+   */
+  public function seek($offset, $whence= SEEK_SET) {
+    $this->file->seek($offset, $whence);
+  }
+
+  /** @return int */
+  public function tell() { return $this->file->tell(); }
 
   /**
    * Truncate this buffer to a given new size.